+++ SpatialGDK/Source/SpatialGDKFunctionalTests/Private/TestMaps/Spatial2WorkerMigrationMaps.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) Improbable Worlds Ltd, All Rights Reserved
 
-#include "TestMaps/Spatial2WorkerFullInterestMigrationMap.h"
+#include "TestMaps/Spatial2WorkerMigrationMaps.h"
 #include "EngineClasses/SpatialWorldSettings.h"
 #include "GameFramework/PlayerStart.h"
 #include "SpatialGDKFunctionalTests/SpatialGDK/SpatialTestCharacterMigration/SpatialTestCharacterMigration.h"
@@ -10,7 +10,6 @@
 USpatial2WorkerFullInterestMigrationMap::USpatial2WorkerFullInterestMigrationMap()
 	: UGeneratedTestMap(EMapCategory::CI_PREMERGE_SPATIAL_ONLY, TEXT("Spatial2WorkerFullInterestMigrationMap"))
 {
-	SetNumberOfClients(2);
 }
 
 void USpatial2WorkerFullInterestMigrationMap::CreateCustomContentForMap()
@@ -29,6 +28,32 @@ void USpatial2WorkerFullInterestMigrationMap::CreateCustomContentForMap()
 	AddActorToLevel<APlayerStart>(CurrentLevel, PlayerTransform);
 
 	ASpatialWorldSettings* WorldSettings = CastChecked<ASpatialWorldSettings>(World->GetWorldSettings());
-	WorldSettings->SetMultiWorkerSettingsClass(UTest2x1FullInterestWorkerSettings::StaticClass());
 	WorldSettings->DefaultGameMode = ACharacterMovementTestGameMode::StaticClass();
+	WorldSettings->SetMultiWorkerSettingsClass(UTest2x1FullInterestWorkerSettings::StaticClass());
 }
+
+USpatial2WorkerNoInterestMigrationMap::USpatial2WorkerNoInterestMigrationMap()
+	: UGeneratedTestMap(EMapCategory::CI_PREMERGE_SPATIAL_ONLY, TEXT("Spatial2WorkerNoInterestMigrationMap"))
+{
+}
+
+void USpatial2WorkerNoInterestMigrationMap::CreateCustomContentForMap()
+{
+	ULevel* CurrentLevel = World->GetCurrentLevel();
+
+	// Add the tests
+	AddActorToLevel<ASpatialTestCharacterMigration>(CurrentLevel, FTransform::Identity);
+
+	FTransform PlayerTransform;
+
+	PlayerTransform.SetTranslation(FVector(-500.0f, -500.0f, 50.0f));
+	AddActorToLevel<APlayerStart>(CurrentLevel, PlayerTransform);
+
+	PlayerTransform.SetTranslation(FVector(-500.0f, 500.0f, 50.0f));
+	AddActorToLevel<APlayerStart>(CurrentLevel, PlayerTransform);
+
+	ASpatialWorldSettings* WorldSettings = CastChecked<ASpatialWorldSettings>(World->GetWorldSettings());
+	WorldSettings->DefaultGameMode = ACharacterMovementTestGameMode::StaticClass();
+	WorldSettings->SetMultiWorkerSettingsClass(UTest2x1NoInterestWorkerSettings::StaticClass());
+}
+
