+++ SpatialGDK/Source/SpatialGDKFunctionalTests/SpatialGDK/SpatialTestCharacterMovement/CharacterMovementTestGameMode.h
@@ -21,5 +21,7 @@ class ACharacterMovementTestGameMode : public AGameModeBase
 
 private:
 	int32 SpawnRequests = 0;
+
+	UPROPERTY()
 	TArray<APlayerStart*> PlayerStarts;
 };
